postgres_to_es/requirements.txt
python-dotenv==0.21.0
psycopg2==2.9.3
elasticsearch==8.4.3
pydantic==1.10.2